mp4-to-gif-converter/desktop_app/requirements.txt
Flask
pywebview
waitress
